Cache worn armour/gizmo egos for faster access
Looking up whether or not the player is wearing particular egos is done
very frequently and involves iterating the player's entire list of equipped
items.

Instead, cache worn egos on equipment change, using the same logic that
artprops already do.

(This is probably of minimal benefit, but should still be slightly faster,
and is potentially another step closer to backend unification of some
egos/artprops.)