Improve interaction of armour egos and randarts, allow some to generate
While there has never been technical reasons why artefact armour cannot
have arbitrary egos (and, in fact, orbs and scarves have done this all the
time), it was only allowed for vault-placed items and sometimes led to
confusing situations for players nonetheless.

For instance, a plate armour of fire resistance could roll an rF+ artprop,
leaving its properties showing as {rF+, rF+}, even though this was
functionally identical to {rF++}. The most common cause of this was likely
vaults specifying an ego manually and also good_item (for random plusses),
which would sometimes create a randart also using the ego.

However, there were always a handful of armour egos that never appeared on
randarts at all (eg: Infusion or Hurling). Recent commits adding even more
of these (and more exciting ones) have had numerous players requesting it
be possible they show up on randarts and so this commit refactors a few
things to allow it to happen.

There is now a function to map ego->artprops, for cases where there is a
one-to-one mapping. In cases where an artefact has a prespecified ego that
*could* be an artprop, it will simply get converted to that artprop upon
generation. (ie: that {rF+, rF+} armour will generate as {rF++} only). If
the ego cannot become an artprop, it will be preserved as an ego.

And for naturally generated artefacts, they have a 50% chance of
generating an ego as normal and then keeping it *only* if it has no artprop
equivalent. (ie: armour will not generate rF+ at a higher rate than before,
but randart armours of Command will be possible).

Hopefully this will allow a few more interesting artefact combinations,
while reducing player confusion at the same time.