Don't give inhibited regen to Armataurs (mysticsailboat)
Apparently it's been possible for Armataurs to be hit with inhibited
regen, preventing regeneration when enemies were visible. Given that
Armataurs were rebalanced around having access to the core gimmick
of rollpage, this was *really* rough for any game hit with it, and
this should have been prevented from the beginning.

Sorry about that!

Here we do save compat in addition to making the muts exclusive,
to help out any affected games in-progress.