Make Orbs of Guile scale with evocations
Guile was by far the most successful of the previous orb designs, being
extremely worth the tradeoff of a shield for specific playstyles. But it
was also arguably *too* powerful - singlehandedly making an entire
repertoir of spells strong the whole game long, reducing many lategame
enemies to 0 will, and generally making it feel night and day whether your
hexer found this specific item.

It is still a cool and exciting effect, and the situations it generates are
interesting, but requiring additional investment may make its power level
more reasonable. So, in keeping with the majority of orbs, it now scales
directly with evocations skill - requiring 16 skill to reach its old level
of power (though also being able to exceed it, for an even higher
investment).