Make Orb of Energy scale with evocations, adjust Wucad Mu
Orbs of energy were a little underwhelming - sometimes useable, but
generally considered not worth abandoning a shield for. Scaling their
effect with evocations may make them more tempting to use. It reaches its
current level of power with 4.5 skill and can greatly exceed it with
significant investment (currently up to 45% chance at 27 evo).

Wucad Mu's effect is currently 'orb of energy x2' and obviously this would
be absurd to double at this point, so I've given it something else. Leaning
into both the monastic flavor of the wizlab, and how the named statue there
drains MP, the crystal ball of Wucad Mu is now an orb of energy that has
a chance to drain additional MP (and briefly antimagic) any hostile
spellcasters in sight whenever the MP refund effect triggers.